refactor: creating package peer inside internal transports p2p
